+++ HISTORY.rst
@@ -151,6 +151,7 @@
 - Split the variable `DOCKER_MYSQL_IMAGE_VERSION` into
   `DOCKER_MYSQL_IMAGE_NAME` and `DOCKER_MYSQL_IMAGE_TAG`.
 - Move the Google credentials backend to its own app.
+- Add support for deleting multiple transformations.
 
 4.9.5 (XXXX-XX-XX)
 ==================